+++ packages/core/src/chat/components/index.ts
@@ -1,0 +1,13 @@
+import ChatInputField from './ChatInputField';
+import ChatInput from './ChatInput';
+import ChatPressableIcon from './ChatPressableIcon';
+import ChatMessage from './ChatMessage';
+import ChatMessageAvatar from './ChatMessageAvatar';
+
+export {
+  ChatInputField,
+  ChatInput,
+  ChatPressableIcon,
+  ChatMessage,
+  ChatMessageAvatar,
+};