+++ packages/core/src/controls/Swipe/types.ts
@@ -1,5 +1,5 @@
 export interface ISwipe {
   active: boolean;
   disabled: boolean;
-  onPress: (event: boolean) => void;
+  onPress: (isActive: boolean) => void;
 }