+++ packages/core/src/other/components/Avatar/AvatarGroup.tsx
@@ -11,9 +11,9 @@ import Avatar from './Avatar';
 import {IAvatarGroupProps} from './types';
 
 const AvatarGroup: FC<IAvatarGroupProps> = props => {
-  const {groups, ...otherProps} = props;
-  const [styles] = useStyles(stylesCreate);
-  const count = groups.length - 3;
+  const {groups, groupCount = groups.length, ...otherProps} = props;
+  const count = groupCount - 3;
+  const [styles] = useStyles(stylesCreate, groupCount);
   const maxCount = 99;
 
   const text = count > maxCount ? maxCount : count;
@@ -45,7 +45,7 @@ const AvatarGroup: FC<IAvatarGroupProps> = props => {
         />
       )}
 
-      {count < 3 ? (
+      {groupCount < 5 ? (
         groups[3] && (
           <Avatar
             user={groups[3]}
@@ -67,10 +67,14 @@ const AvatarGroup: FC<IAvatarGroupProps> = props => {
 
 export default AvatarGroup;
 
-const stylesCreate = ({colors, spaces}: IThemeContext) => {
+const stylesCreate = ({colors, spaces}: IThemeContext, length: number) => {
   return StyleSheet.create({
     container: {
       flexDirection: 'row',
+      maxWidth:
+        length > 3
+          ? spaces.Space40 * 4 - rem(36)
+          : spaces.Space40 * length - spaces.Space12 * (length - 1),
     },
     countView: {
       width: spaces.Space40,
@@ -79,6 +83,7 @@ const stylesCreate = ({colors, spaces}: IThemeContext) => {
       backgroundColor: colors.BgSecondary,
       justifyContent: 'center',
       zIndex: 4,
+      right: rem(36),
 
       borderWidth: spaces.Space2,
       borderColor: colors.BgPrimary,
@@ -88,18 +93,19 @@ const stylesCreate = ({colors, spaces}: IThemeContext) => {
       textAlign: 'center',
     },
     avatarOne: {
-      left: rem(36),
       zIndex: 1,
     },
     avatarTwo: {
-      left: spaces.Space24,
+      right: spaces.Space12,
+
       zIndex: 2,
     },
     avatarThree: {
-      left: spaces.Space12,
+      right: spaces.Space24,
       zIndex: 3,
     },
     avatarFour: {
+      right: rem(36),
       zIndex: 4,
     },
   });