+++ packages/core/src/other/components/Carousel/index.ts
@@ -1,4 +1,5 @@
 import Carousel from './Carousel';
+import LoopCarousel from './LoopCarousel';
 
-export {Carousel};
+export {Carousel, LoopCarousel};
 export * from './types';